move each allocation row into its own component
